requirements.txt
google-genai
python-dotenv
langchain
streamlit
langchain-core
langchain-groq

